refactor(front): move slack credentials DTO out of lib/api into types/api
The mirror target types/api/credentials.ts is already an unrelated module
(provider/dust-managed credentials helpers), so the Slack-legacy DTO gets a
distinct, route-mirroring home.

- Moved GetSlackClientIdResponseBody from lib/api/credentials.ts to
  types/api/credentials/slack_is_legacy.ts (matches the
  /credentials/slack_is_legacy route). lib/api/credentials.ts deleted.
- Updated 2 consumers (front/lib/swr/oauth.ts, the slack_is_legacy route).
