+++ front/lib/swr/oauth.ts
@@ -1,6 +1,6 @@
-import type { GetSlackClientIdResponseBody } from "@app/lib/api/credentials";
 import { clientFetch } from "@app/lib/egress/client";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
+import type { GetSlackClientIdResponseBody } from "@app/types/api/credentials/slack_is_legacy";
 import type { GetOAuthSetupResponseBody } from "@app/types/api/oauth";
 import type { APIError, WithAPIErrorResponse } from "@app/types/error";
 import { isAPIErrorResponse } from "@app/types/error";
